로그인 확인 과정, 주문 상태 커밋

+++ server/graphql/resolvers/index.js
@@ -5,8 +5,9 @@ import jsonwebtoken from 'jsonwebtoken';
 // resolver에서 mutation을 정의하고 구현하는 걸 보니 가장 중요한 부분이 아닐까 싶다. service 단이라고 생각하자
 const resolvers = {
     Query: {
-        async orders(_, args) {
+        async orders(_, args,{user}) {
             try {
+                if (!user) throw new Error('You are not authenticated')
                 let orders = await Order.find().sort({createdAt: -1});
                 const search = args.search || "";
                 const category = args.category;
@@ -114,27 +115,34 @@ const resolvers = {
         }
     },
     Mutation: {
-        createOrder: async (_, args) => {
+        createOrder: async (_, args, {user}) => {
             try {
+                if(!user) throw error("로그인 되어 있지 않습니다.");
                 const order = new Order({
                     ...args.orderInput
                 })
+                
+                await users.findOneAndUpdate(user._id,{status:"주문완료"});
                 const result = await order.save();
                 return result;
             } catch (e) {
                 throw new Error('Error: ', e);
             }
         },
-        removeOrder: async (_, args) => {
+        removeOrder: async (_, args,{user}) => {
             try {
+                if(!user) throw error("로그인 되어 있지 않습니다.");
+                await users.findOneAndUpdate(user._id,{status:"주문취소"});
                 const removedorder = await Order.findByIdAndRemove(args._id).exec()
                 return removedorder
             } catch (e) {
                 throw new Error('Error: ', e)
             }
         },
-        updateOrder: async (_, {_id, menu, hi}) => {
+        updateOrder: async (_, {_id, menu, hi},{user}) => {
             try {
+                if(!user) throw error("로그인 되어 있지 않습니다.");
+                await users.findOneAndUpdate(user._id,{status:"주문완료"});
                 const updatedOrder = await Order.findByIdAndUpdate(_id, {
                     $set: {menu, hi}
                 }).exec()
@@ -142,6 +150,11 @@ const resolvers = {
             } catch (e) {
                 throw new Error('Error: ', e)
             }
+        },
+        giveupOrder: async (_, args,{user})=>{
+            if(!user) throw error("로그인 되어 있지 않습니다.");
+            await users.findOneAndUpdate(user._id,{status:"주문포기"});
+            return "주문을 포기하셨습니다."
         },
         searchByID: async (_, args) => {
             try {
